delete validation

+++ dd_roller.py
@@ -9,7 +9,6 @@ from .db import get_db
 bp = Blueprint('dd_roller', __name__)
 
 
-
 @bp.route('/roll')
 @login_required
 def roll():
@@ -18,8 +17,9 @@ def roll():
         'SELECT c.id, character_name, character_class, race, strenght, dexterity, constitution, intelligence, wisdom, charisma, created, author_id, username'
         ' FROM character c JOIN user u ON c.author_id = u.id'
         ' ORDER BY created DESC'
-    ).fetchall() 
+    ).fetchall()
     return render_template('dd_roller/roll.html', characters=characters)
+
 
 @bp.route('/character', methods=('GET', 'POST'))
 @login_required
@@ -39,26 +39,26 @@ def character():
         if not name:
             error = 'Name is required.'
         if not strenght:
-            error = 'Attributes are required.'    
+            error = 'Attributes are required.'
 
         if error is not None:
             flash(error)
         else:
             db = get_db()
             db.execute(
-                'INSERT INTO character (character_name, character_class, race, strenght, dexterity, constitution, intelligence, wisdom, charisma, author_id)' #columnes in data base 
+                # columnes in data base
+                'INSERT INTO character (character_name, character_class, race, strenght, dexterity, constitution, intelligence, wisdom, charisma, author_id)'
                 ' VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
-                (name, class_character, race, strenght, dexterity, constitution, intelligence, wisdom, charisma, g.user['id']) 
+                (name, class_character, race, strenght, dexterity,
+                 constitution, intelligence, wisdom, charisma, g.user['id'])
             )
             db.commit()
         return redirect(url_for('roll'))
-    
-    return render_template('dd_roller/create_character.html')  
+
+    return render_template('dd_roller/create_character.html')
 
 
-
-
-def get_character(id, check_author=True):                                 #checking if author of post is logged in
+def get_character(id, check_author=True):  # checking if author of post is logged in
     character = get_db().execute(
         'SELECT c.id, character_name, character_class, race, strenght, dexterity, constitution, intelligence, wisdom, charisma, author_id'
         ' FROM character c JOIN user u ON c.author_id = u.id'
@@ -79,8 +79,13 @@ def get_character(id, check_author=True):                                 #check
 @login_required
 def delete():
     id = request.form['id']
-    db = get_db()
-    db.execute('DELETE FROM character WHERE id = ?', (id,))
-    db.commit()
-    flash('Character deleted')
+    character = get_character(id, check_author=False)
+    
+    if character['author_id'] != g.user['id']:
+        flash('This is not your character')
+    else:    
+        db = get_db()
+        db.execute('DELETE FROM character WHERE id = ?', (id,))
+        db.commit()
+        flash('Character deleted')
     return redirect(url_for('roll'))
